Fixed demo

+++ data/shaders/test.vert
@@ -6,7 +6,7 @@ in vec2 UV;
 out vec2 vert_uv;
 
 uniform vec2 uv_offset;
-uniform vec2 translate;
+uniform ivec2 translate;
 uniform vec2 scale;
 uniform vec2 cam_position;
 uniform ivec2 cam_resolution;
